Request state machine, clippy fixes, CI protoc
Request state machine:
- RequestTracker with 4 states (Idle, Streaming, Completed, Errored)
- Sendfile failure now sends Error frame instead of silent truncation
- Scan streaming sends Error frame on mid-stream write failure
- BatchRead write failure returns Error instead of empty response

Clippy fixes:
- Remove unused imports (BufMut, Hasher, CACHE_PUT_ERRORS)
- Remove dead fields (max_inflight_bytes, sendfile_timeout_secs)
- Use BuildHasher::hash_one instead of manual hasher
- Use Range::contains instead of manual comparison
- Use Option::map instead of if-let-else
- Add Default impls for SpscRing and ThreadChannel
- Fix loop indexing with enumerate
- Allow dead_code on utility functions (len, check_disk_space)
- Prefix unused fields with underscore (_vnodes, _config, _ready)

CI:
- Install protobuf-compiler for etcd-client build
- Remove external project references from comments

+++ bench/benches/eviction_policy.rs
@@ -51,7 +51,7 @@ fn scan_resistant_sequence(hot_size: u64, total_keys: u64, length: usize, seed: 
 
         if r < 0.8 {
             rng = rng.wrapping_mul(6364136223846793005).wrapping_add(1);
-            let idx = (rng >> 33) as u64 % hot_size;
+            let idx = (rng >> 33) % hot_size;
             result.push(idx);
         } else {
             result.push(scan_pos);
